// generated/src/app/api-models/commission-determ-y.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CommissionDetermY {
  matrixDeterminant: string;
  matrixSeq: number;
  yAxisDescription: string;
  premiumPaidYtdFrom: number;
  premiumPaidYtdThru: number;
  premiumBilledYtdFrom: number;
  premiumBilledYtdThru: number;
  state: string;
  lineOfBusiness: string;
  yearsContractedFrom: number;
  yearsContractedThru: number;
  yearsPeriod: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}